feat: Add support for sending documents in TelegramBot
This commit adds a new method, SendDocumentToStorage, to handle sending documents in the TelegramBot. It checks the MIME type of the document and saves it to the appropriate storage container. If the document has a thumbnail, it also saves the thumbnail to the "thumbnails" container. This feature enhances the functionality of the TelegramBot by allowing users to send and store various types of documents.